requirements.txt
annotated-types==0.7.0
black==24.10.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comtypes==1.4.8
confection==0.1.5
cymem==2.0.10
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
mypy-extensions==1.0.0
numpy==2.0.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
preshed==3.0.9
pvporcupine==3.0.3
PyAudio==0.2.14
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.98
pywin32==308
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.6.0
shellingham==1.5.4
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.11.0
srsly==2.4.8
thinc==8.3.2
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.0
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0